add:宏控制截屏方式
