+++ app/build.gradle
@@ -1,16 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
+    signingConfigs {
+    }
     compileSdkVersion 22
     buildToolsVersion '22.0.1'
-
     defaultConfig {
         applicationId 'com.intfocus.yonghuitest'
         minSdkVersion 14
         targetSdkVersion 22
     }
-
-    if(System.getenv().containsKey("KEYSTORE"))
+    if (System.getenv().containsKey("KEYSTORE"))
         signingConfigs {
             release {
                 storeFile file(System.getenv("KEYSTORE"))
@@ -19,17 +19,14 @@ android {
                 keyPassword System.getenv("KEY_PASSWORD")
             }
         }
-
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles 'proguard-rules.pro' //getDefaultProguardFile('proguard-android.txt'),
-
-            if(System.getenv().containsKey("KEYSTORE"))
+            if (System.getenv().containsKey("KEYSTORE"))
                 signingConfig signingConfigs.release
         }
     }
-
     lintOptions {
         abortOnError false
         checkReleaseBuilds false
@@ -47,7 +44,7 @@ dependencies {
     // compile 'com.joanzapata.pdfview:android-pdfview:1.0.4@aar'
     compile 'com.android.support:appcompat-v7:22.1.0'
     compile 'com.android.support:support-v4:22.1.0'
-    compile ('com.pgyersdk:sdk:2.2.2')
+    compile 'com.pgyersdk:sdk:2.2.2'
     compile 'commons-io:commons-io:+'
     compile 'com.squareup.okhttp3:okhttp:3.2.0'
     compile 'me.dm7.barcodescanner:zbar:1.8.4'
