+++ app/src/main/java/com/intfocus/yh_android/util/URLs.java
@@ -70,6 +70,7 @@ public class URLs implements Serializable {
     public final static String kETag                     = "ETag";
     public final static String kLastModified             = "Last_Modified";
     public final static boolean kIsQRCode                = true;
+    public final static boolean kIsFullScreen            = true;
 
     public static String storage_base(Context context) {
         //    String path = "";
